+++ src/pages/subjects/[subject].tsx
@@ -12,7 +12,7 @@ import { useRouter } from "next/router";
 import Offline from "../../components/Offline";
 import { offlineContext } from "../_app";
 import NavTabs from "./Tabs";
-import Drawer from "./allDrawer";
+import Drawer from "./AllDrawer";
 
 // import TabsPC from "./TabsPc";
 // import TabsPhone from "./TabsPhone";
